requirements.txt
streamlit
numpy
scikit-learn
joblib
opencv-python-headless
Pillow
